Finished Marking RM Tree Red
This is another branch finishing the marking of RM Tree in red. The
previous branch seems weird this day. The content of
markedRTreeData.csv has ‘red’ at the value field, but the data in the
d3.csv(“static/markedRTreeData.csv”, function(error, data){} has
‘black’ at the value field. This problem was solved by rename the
markedRTreeData.csv to a different name markedRTree.csv. The reason is
not clear yet.